requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
asyncpg==0.29.0

# Message Queue
kafka-python==2.0.2
celery==5.3.4

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.4
pandas==2.1.4
joblib==1.3.2

# Technical Analysis
ta-lib==0.4.28
yfinance==0.2.28
ccxt==4.1.64

# Data Processing
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# API Clients
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
python-binance==1.0.19
alpaca-trade-api==3.1.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
loguru==0.7.2

# Security
cryptography==41.0.8
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
supervisor==4.2.5
